+++ app/src/main/java/com/yanbin/reactivestickynote/di/NoteModule.kt
@@ -2,13 +2,12 @@ package com.yanbin.reactivestickynote.di
 
 import com.yanbin.reactivestickynote.account.AccountService
 import com.yanbin.reactivestickynote.account.SPAccountService
-import com.yanbin.reactivestickynote.editor.data.FirebaseNoteRepository
-import com.yanbin.reactivestickynote.editor.data.NoteRepository
+import com.yanbin.reactivestickynote.stickynote.data.FirebaseNoteRepository
+import com.yanbin.reactivestickynote.stickynote.data.NoteRepository
 import com.yanbin.reactivestickynote.editor.vm.ContextMenuViewModel
 import com.yanbin.reactivestickynote.edittext.EditTextViewModel
 import com.yanbin.reactivestickynote.editor.vm.EditorViewModel
 import com.yanbin.reactivestickynote.editor.domain.StickyNoteEditor
-import com.yanbin.reactivestickynote.editor.domain.ViewPort
 import com.yanbin.reactivestickynote.login.LoginViewModel
 import com.yanbin.reactivestickynote.editor.vm.StickyNoteViewModel
 import com.yanbin.reactivestickynote.editor.vm.ViewPortViewModel
