+++ app/src/main/java/com/yanbin/reactivestickynote/editor/domain/ViewPort.kt
@@ -1,7 +1,7 @@
 package com.yanbin.reactivestickynote.editor.domain
 
-import com.yanbin.reactivestickynote.editor.data.NoteRepository
-import com.yanbin.reactivestickynote.editor.model.Position
+import com.yanbin.reactivestickynote.stickynote.data.NoteRepository
+import com.yanbin.reactivestickynote.stickynote.model.Position
 import io.reactivex.rxjava3.core.Observable
 import io.reactivex.rxjava3.subjects.BehaviorSubject
