+++ app/src/main/java/com/yanbin/reactivestickynote/editor/usecase/ResizeNoteUseCase.kt
@@ -1,8 +1,8 @@
 package com.yanbin.reactivestickynote.editor.usecase
 
 import com.yanbin.reactivestickynote.editor.domain.StickyNoteEditor
-import com.yanbin.reactivestickynote.editor.model.StickyNote
-import com.yanbin.reactivestickynote.editor.model.YBSize
+import com.yanbin.reactivestickynote.stickynote.model.StickyNote
+import com.yanbin.reactivestickynote.stickynote.model.YBSize
 import io.reactivex.rxjava3.core.Observable
 import io.reactivex.rxjava3.kotlin.addTo
