+++ app/src/main/java/com/yanbin/reactivestickynote/editor/vm/ViewPortViewModel.kt
@@ -2,7 +2,7 @@ package com.yanbin.reactivestickynote.editor.vm
 
 import androidx.lifecycle.ViewModel
 import com.yanbin.reactivestickynote.editor.domain.ViewPort
-import com.yanbin.reactivestickynote.editor.model.Position
+import com.yanbin.reactivestickynote.stickynote.model.Position
 
 class ViewPortViewModel(
     private val viewPort: ViewPort
