+++ app/src/main/java/com/yanbin/reactivestickynote/edittext/EditTextViewModel.kt
@@ -1,7 +1,7 @@
 package com.yanbin.reactivestickynote.edittext
 
 import androidx.lifecycle.ViewModel
-import com.yanbin.reactivestickynote.editor.data.NoteRepository
+import com.yanbin.reactivestickynote.stickynote.data.NoteRepository
 import io.reactivex.rxjava3.core.Observable
 import io.reactivex.rxjava3.disposables.CompositeDisposable
 import io.reactivex.rxjava3.kotlin.addTo
