+++ app/src/main/java/com/yanbin/reactivestickynote/stickynote/data/NoteRepository.kt
@@ -1,8 +1,8 @@
-package com.yanbin.reactivestickynote.editor.data
+package com.yanbin.reactivestickynote.stickynote.data
 
 import com.yanbin.reactivestickynote.account.Account
-import com.yanbin.reactivestickynote.editor.model.SelectedNote
-import com.yanbin.reactivestickynote.editor.model.StickyNote
+import com.yanbin.reactivestickynote.stickynote.model.SelectedNote
+import com.yanbin.reactivestickynote.stickynote.model.StickyNote
 import io.reactivex.rxjava3.core.Observable
 
 interface NoteRepository {
