+++ app/src/main/java/com/yanbin/reactivestickynote/stickynote/model/YBColor.kt
@@ -1,4 +1,4 @@
-package com.yanbin.reactivestickynote.editor.model
+package com.yanbin.reactivestickynote.stickynote.model
 
 data class YBColor(
     val color: Long
