+++ app/src/test/kotlin/com/yanbin/reactivestickynote/editor/domain/EditorViewModelTest.kt
@@ -1,9 +1,9 @@
 package com.yanbin.reactivestickynote.editor.domain
 
-import com.yanbin.reactivestickynote.editor.data.NoteRepository
-import com.yanbin.reactivestickynote.editor.model.Position
-import com.yanbin.reactivestickynote.editor.model.StickyNote
-import com.yanbin.reactivestickynote.editor.model.YBColor
+import com.yanbin.reactivestickynote.stickynote.data.NoteRepository
+import com.yanbin.reactivestickynote.stickynote.model.Position
+import com.yanbin.reactivestickynote.stickynote.model.StickyNote
+import com.yanbin.reactivestickynote.stickynote.model.YBColor
 import io.mockk.every
 import io.mockk.mockk
 import io.mockk.verify
